requirements.txt
# Зависимости для проекта ассемблера и интерпретатора УВМ
# Python 3.7+ требуется

# Для парсинга YAML файлов
PyYAML>=6.0

